mngr_gcp: move GCS fakes/stubs to testing.py
Problem: The in-memory GCS fakes (``_FakeBlob`` / ``_FakeBucket`` /
``_FakeIamConfiguration`` / ``_FakeListIterator`` / ``_FakeStorageClient``)
and the test-only injection seams (``_StubbedGcsStateBucket`` /
``_StubbedGcsVolume``) plus the shared ``_FAKE_CREDENTIALS`` constant lived in
``state_bucket_test.py`` and were imported from ``state_bucket_backend_test.py``.
The repo convention is that ``testing.py`` houses cross-test fakes and stubs:
``libs/mngr_gcp/imbue/mngr_gcp/testing.py`` already houses
``FakeInstancesClient`` / ``FakeFirewallsClient`` / ``_StubbedGcpVpsClient``,
and ``libs/mngr_azure/imbue/mngr_azure/testing.py`` houses
``_StubbedBlobStateBucket`` for the same reason.

Fix: Lift the fakes / stubs / ``_FAKE_CREDENTIALS`` into
``mngr_gcp/testing.py`` and update both test files to import from there.
``state_bucket_test.py`` keeps its file-local factory helpers (``_make_bucket``
/ ``_make_volume``) and the ``_FlakyLookupClient`` subclass that's only used
by one test.
